backend/pyproject.toml
[tool.poetry]
name = "ai-agency-saas"
version = "0.1.0"
description = "AI Agency SaaS Platform"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.0"
sqlalchemy = "^2.0.0"
pydantic = "^2.0.0"
pydantic-settings = "^2.0.0"
celery = "^5.3.0"
redis = "^5.0.0"
diffusers = "^0.24.0"
transformers = "^4.35.0"
torch = "^2.1.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
psycopg2-binary = "^2.9.9"
alembic = "^1.12.0"
boto3 = "^1.34.0"
pillow = "^10.0.0" 